code/src/cloud_functions/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-monitoring
scikit-learn
